Update profilepic for User

+++ resources/views/message.blade.php
@@ -1,9 +1,19 @@
 @extends('layouts.app')
 @section('content')
+@if (session('success'))
+    <div class="alert alert-success">
+        {{ session('success') }}
+    </div>
+@elseif (session('error'))
+    <div class="alert alert-danger">
+        {{ session('error') }}
+    </div>
+@endif
     <table class="table table-striped">
         <thead>
             <tr>
                 <th>ID</th>
+                <th>auteur</th>
                 <th>Content</th>
                 <th></th>
                 <th>Created at</th>
@@ -15,11 +25,20 @@
     @foreach($messages as $message)
             <tr>
                 <td>{{$message->id}}</td>
+                <td>{{$message->author}}</td>
                 <td class="content">{{$message->content}}</td>
-                <td><a href="/edit/{{$message->id}}">edit</a></td>
-                <td>{{$message->created_at}}</td>
-                <td>{{$message->updated_at}}</td>
-                <td><a href="/delete/{{$message->id}}">X</a></td>
+                @if ( Auth::user() && $message->author == Auth::user()->name)
+                            <td><a href="/edit/{{$message->id}}">edit</a></td>
+                            <td>{{$message->created_at}}</td>
+                            <td>{{$message->updated_at}}</td>
+                            <td><a href="/delete/{{$message->id}}">X</a></td>
+                @else
+                            <td></td>
+                            <td>{{$message->created_at}}</td>
+                            <td>{{$message->updated_at}}</td>
+                            <td></td>
+                
+                @endif
             </tr>
     @endforeach
         </tbody>
